исправлены замечания. изменена структура файлов.
