+++ .gitignore
@@ -1,0 +1,2 @@
+/ml_project/data/raw/gp_kazan.jpg
+/ml_project/src/model/outputs/
