+++ ml_project/src/model/conf/config.yaml
@@ -9,5 +9,5 @@ debug_image_dir: c:/temp/
 batch_size: 256
 threshold: 0.95
 # путь к корневому каталогу ПРОГРАММЫ
-path_to_root: ../../../../..
+path_to_root: ../../
