+++ ml_project/src/model/conf/train_config.yaml
@@ -2,11 +2,11 @@ defaults:
   - map: kazan_cfg.yaml
 
   # путь к корневому каталогу
-path_to_root: ..\..\..\..\..
+path_to_root: ../../
   # путь от корневого каталога до изображения карты
 
 walks_file: data\interim\walks4.txt
-model_save: best_model_crop32.pkl
+model_save: model\best_model_crop32.pkl
 train_val_ratio: 0.8
 batch_size: 512
 crop_size: 32
