+++ ml_project/src/model/conf/train_config2.yaml
@@ -2,7 +2,7 @@ defaults:
   - map: kazan_cfg.yaml
 
   # путь к корневому каталогу
-path_to_root: ..\..\..\..\..
+path_to_root: ../../
   # путь от корневого каталога до изображения карты
