+++ setup.py
@@ -1,0 +1,12 @@
+from setuptools import setup
+
+setup(
+    name='geomap6',
+    version='6.2.0',
+    packages=['ml_project', 'ml_project.src', 'ml_project.src.model', 'ml_project.src.model.utils'],
+    url='',
+    license='MIT',
+    author='asokolov',
+    author_email='aesokolov1975@gmail.com',
+    description='Geo tools'
+)
